TINY-8391: Move 'InsertTable' command to model
