+++ modules/tinymce/src/core/main/ts/api/OptionTypes.ts
@@ -49,10 +49,6 @@ export interface ToolbarGroup {
 
 export type ToolbarMode = 'floating' | 'sliding' | 'scrolling' | 'wrap';
 
-export type TableSizingMode = 'fixed' | 'relative' | 'responsive' | 'auto';
-export type TableColumnResizing = 'preservetable' | 'resizetable';
-export type TableHeaderType = 'section' | 'cells' | 'sectionCells' | 'auto';
-
 interface BaseEditorOptions {
   a11y_advanced_options?: boolean;
   add_form_submit_trigger?: boolean;
@@ -199,16 +195,7 @@ interface BaseEditorOptions {
   style_formats_merge?: boolean;
   submit_patch?: boolean;
   suffix?: string;
-  // TODO: Remove table options that are no longer used
-  table_clone_elements?: string[];
-  table_column_resizing?: TableColumnResizing;
-  table_default_attributes?: Record<string, string>;
-  table_default_styles?: Record<string, string>;
-  table_header_type?: TableHeaderType;
-  table_resize_bars?: boolean;
-  table_sizing_mode?: TableSizingMode;
   table_tab_navigation?: boolean;
-  table_use_colgroups?: boolean;
   target?: HTMLElement;
   text_patterns?: RawPattern[] | false;
   theme?: string | ThemeInitFunc | false;
