+++ modules/tinymce/src/core/main/ts/api/commands/Commands.ts
@@ -17,7 +17,6 @@ import * as LinkCommands from './LinkCommands';
 import * as ListCommands from './ListCommands';
 import * as NewlineCommands from './NewlineCommands';
 import * as SelectionCommands from './SelectionCommands';
-import * as TableCommands from './TableCommands';
 
 const registerExecCommands = (editor: Editor): void => {
   editor.editorCommands.addCommands({
@@ -58,7 +57,6 @@ export const registerCommands = (editor: Editor): void => {
   NewlineCommands.registerCommands(editor);
   ListCommands.registerCommands(editor);
   FormatCommands.registerCommands(editor);
-  TableCommands.registerCommands(editor);
 
   registerExecCommands(editor);
 };
