+++ modules/tinymce/src/core/main/ts/table/TableUtils.ts
@@ -13,9 +13,7 @@
  Make sure that if making changes to this file, the other files are updated as well
  */
 
-import { Arr } from '@ephox/katamari';
-import { TableLookup } from '@ephox/snooker';
-import { Attribute, Compare, SugarElement } from '@ephox/sugar';
+import { Compare, SugarElement } from '@ephox/sugar';
 
 import Editor from '../api/Editor';
 
@@ -25,24 +23,7 @@ const getBody = (editor: Editor): SugarElement<HTMLElement> =>
 const getIsRoot = (editor: Editor) => (element: SugarElement<Node>): boolean =>
   Compare.eq(element, getBody(editor));
 
-const removeDataStyle = (table: SugarElement<HTMLTableElement>): void => {
-  Attribute.remove(table, 'data-mce-style');
-
-  const removeStyleAttribute = (element: SugarElement<HTMLElement>) => Attribute.remove(element, 'data-mce-style');
-
-  Arr.each(TableLookup.cells(table), removeStyleAttribute);
-  Arr.each(TableLookup.columns(table), removeStyleAttribute);
-  Arr.each(TableLookup.rows(table), removeStyleAttribute);
-};
-
-// TODO: Remove these and move TableUtilsTest.ts to model as well
-const isPercentage = (value: string): boolean => /^(\d+(\.\d+)?)%$/.test(value);
-const isPixel = (value: string): boolean => /^(\d+(\.\d+)?)px$/.test(value);
-
 export {
   getBody,
-  getIsRoot,
-  removeDataStyle,
-  isPercentage,
-  isPixel
+  getIsRoot
 };
