+++ modules/tinymce/src/models/dom/main/ts/table/api/Options.ts
@@ -5,10 +5,15 @@
  * For commercial licenses see https://www.tiny.cloud/
  */
 
-import { Arr } from '@ephox/katamari';
+import { Arr, Obj } from '@ephox/katamari';
+import { SugarElement, Width } from '@ephox/sugar';
 
 import Editor from 'tinymce/core/api/Editor';
 import { EditorOptions } from 'tinymce/core/api/OptionTypes';
+
+export type TableSizingMode = 'fixed' | 'relative' | 'responsive' | 'auto';
+export type TableColumnResizing = 'preservetable' | 'resizetable';
+export type TableHeaderType = 'section' | 'cells' | 'sectionCells' | 'auto';
 
 const option: {
   <K extends keyof EditorOptions>(name: K): (editor: Editor) => EditorOptions[K] | undefined;
@@ -16,11 +21,62 @@ const option: {
 } = (name: string) => (editor: Editor) =>
   editor.options.get(name);
 
+// Note: This is also specified in the table plugin Options.ts file
+const defaultTableStyles = {
+  'border-collapse': 'collapse',
+  'width': '100%'
+};
+
+// Note: This is also contained in the table plugin Options.ts file
+const determineDefaultTableStyles = (editor: Editor): Record<string, string> => {
+  if (isTablePixelsForced(editor)) {
+    // Determine the inner size of the parent block element where the table will be inserted
+    const dom = editor.dom;
+    const parentBlock = dom.getParent<HTMLElement>(editor.selection.getStart(), dom.isBlock) ?? editor.getBody();
+    const contentWidth = Width.getInner(SugarElement.fromDom(parentBlock));
+    return { ...defaultTableStyles, width: contentWidth + 'px' };
+  } else if (isTableResponsiveForced(editor)) {
+    return Obj.filter(defaultTableStyles, (_value, key) => key !== 'width');
+  } else {
+    return defaultTableStyles;
+  }
+};
+
 const register = (editor: Editor): void => {
   const registerOption = editor.options.register;
 
   registerOption('table_clone_elements', {
     processor: 'string[]'
+  });
+
+  registerOption('table_use_colgroups', {
+    processor: 'boolean',
+    default: false
+  });
+
+  registerOption('table_header_type', {
+    processor: (value) => {
+      const valid = Arr.contains([ 'section', 'cells', 'sectionCells', 'auto' ], value);
+      return valid ? { value, valid } : { valid: false, message: 'Must be one of: section, cells, sectionCells or auto.' };
+    },
+    default: 'section'
+  });
+
+  registerOption('table_sizing_mode', {
+    processor: 'string',
+    default: 'auto'
+  });
+
+  registerOption('table_default_attributes', {
+    processor: 'object',
+    default: {
+      border: '1'
+    }
+  });
+
+  registerOption('table_default_styles', {
+    processor: 'object',
+    default: defaultTableStyles
   });
 
   registerOption('table_column_resizing', {
@@ -63,6 +119,16 @@ const hasTableObjectResizing = (editor: Editor): boolean => {
   return Arr.contains(objectResizing.split(','), 'table');
 };
 
+const getTableDefaultAttributes = option('table_default_attributes');
+
+const getTableDefaultStyles = (editor: Editor): Record<string, string> => {
+  // Note: The we don't rely on the default here as we need to dynamically lookup the widths based on the current editor state
+  const options = editor.options;
+  return options.isSet('table_default_styles') ? options.get('table_default_styles') : determineDefaultTableStyles(editor);
+};
+
+const tableUseColumnGroup = option('table_use_colgroups');
+
 export {
   register,
 
@@ -74,5 +140,8 @@ export {
   getColumnResizingBehaviour,
   getTableColumnResizingBehaviour,
   hasTableObjectResizing,
-  hasTableResizeBars
+  hasTableResizeBars,
+  getTableDefaultAttributes,
+  getTableDefaultStyles,
+  tableUseColumnGroup
 };
